+++ database/migrations/2021_11_22_151250_create_required_items_table.php
@@ -17,7 +17,7 @@ class CreateRequiredItemsTable extends Migration
             $table->id();
             $table-> string('requiredItemId');
             $table-> string('requiredItem');
-            $table-> int('requiredAmount');
+            $table-> unsignedBigInteger('requiredAmount');
             $table-> double('projectedCost');
             $table-> string('requiredStatus');
             $table->timestamps();
