+++ database/migrations/2021_11_22_153042_create_order_items_table.php
@@ -17,7 +17,7 @@ class CreateOrderItemsTable extends Migration
             $table->id();
             $table-> string('itemId');
             $table-> string('itemName') ;
-            $table-> int('itemOrderAmount');
+            $table-> unsignedBigInteger('itemOrderAmount');
             $table-> string('orderId');
             $table-> string('paymentStatus');
             $table->timestamps();
