+++ database/migrations/2021_11_22_153141_create_orders_table.php
@@ -16,7 +16,7 @@ class CreateOrdersTable extends Migration
         Schema::create('orders', function (Blueprint $table) {
             $table->id();
             $table-> string('orderId');
-            $table-> int('orderItemsCount');
+            $table-> unsignedBigInteger('orderItemsCount');
             $table-> string('orderDate');
             $table-> string('paymentStatus');
             $table-> string('retailNameId');
