+++ database/migrations/2021_11_22_153318_create_expenses_table.php
@@ -20,7 +20,7 @@ class CreateExpensesTable extends Migration
             $table-> string('dayOfWeek');
             $table-> double('cost');
             $table-> string('month');
-            $table-> int('year');
+            $table-> unsignedBigInteger('year');
             $table->timestamps();
         });
     }
