+++ database/migrations/2021_11_22_153336_create_revenues_table.php
@@ -23,7 +23,7 @@ class CreateRevenuesTable extends Migration
             $table-> double('itemPrice');
             $table-> double('sellingPrice');
             $table-> string('month');
-            $table-> int('year');
+            $table-> unsignedBigInteger('year');
             $table->timestamps();
