+++ database/migrations/2021_11_22_153413_create_profits_table.php
@@ -20,7 +20,7 @@ class CreateProfitsTable extends Migration
             $table-> string('dayOfWeek');
             $table-> double('itemsProfit');
             $table-> string('month');
-            $table-> int('year');
+            $table-> unsignedBigInteger('year');
             $table->timestamps();
         });
     }
